Changed all unsigned long values to simply   unsigned to work with our 4 byte encoding.

+++ lpm.c
@@ -14,21 +14,21 @@ void pci_card_init(struct pci_card *card){
 	return;
 }
 
-unsigned long pci_card_read(char* fname, char* raddr){
+unsigned pci_card_read(char* fname, char* raddr){
 	struct pci_card fusion;
 	fusion.filename = fname;
 	fusion.target = strtoul(raddr, 0, 0);
 	pci_card_init(&fusion);
-	fusion.read_result = *((unsigned long *) fusion.virt_addr);
+	fusion.read_result = *((unsigned *) fusion.virt_addr);
 	return fusion.read_result;
 }
 
-void pci_card_write(char* fname, char* raddr, unsigned long wval){
+void pci_card_write(char* fname, char* raddr, unsigned wval){
 	struct pci_card fusion;
 	fusion.filename = fname;
 	fusion.target = strtoul(raddr, 0, 0);
 	pci_card_init(&fusion);
-	*((unsigned long *) fusion.virt_addr) = wval;
+	*((unsigned *) fusion.virt_addr) = wval;
 	return;
 }
